Avoid overflow on integer multiplication
Multiplying any unsigned integer amount by a rate in [0,1] must yield
a product that, when rounded, is representable as that integer type.
Now define a limit that is 1e8 * (maximum value for that integer type).
If amount is less than that limit, then 1e8 * amount cannot overflow;
neither can any lower multiple of the amount. Therefore,
  amount * (1e8 * rate) [amount <= limit]
cannot overflow.

Suppose rate is given as a decimal with no more than eight significant
digits, 0.12345678 for example. If that's available as a double, then
it's an inexact representation of the intended rational 12345678 / 1e8,
but rint(1e8 * rate) recovers the exact numerator of that rational.
Multiplying amount by that numerator gives 1e8 times the exact product.
Dividing the product by 1e8 gives the product; if the remainder is zero,
then it is exact; otherwise, it's as exact as possible, and generally
more accurate than multiplying an integral amount by a double rate
because the double's representation error has been removed. The accuracy
of integer multiplication has been achieved, to the extent possible,
while avoiding overflow.

If this could be explained in far fewer words, it'd be obvious.

This causes a two-cent regression in guaranteed premium on a single
testcase, for the reason explained in the preceding commit.