Remove unneeded try/catch from IllustrationView::OnCreate()
This was a workaround for a problem in wxWidgets fixed a long time ago,
see wxWidgets commit 7047d7981f (handle exceptions thrown from
overridden wxView::OnCreate() gracefully, 2008-12-30) and is not needed
any longer: no view is created if an exception is thrown from here.

This commit is best viewed ignoring whitespace-only changes.