Fix math_functions_test build when using autotools
Put fdlibm_*.c files in SOURCES, not LDADD, as this is what they are.

Note that we don't need to use the hack with the per-target custom
CXXFLAGS here as long as these files are only compiled as part of this
test -- but it would become needed if they were reused in another test
or as part of some library in the future.