Rename fdlibm header
The concinnity test now passes.

Arguably the file extension should be '.h' rather than '.hpp'. The same
could be said, though, of 'config.hpp', which is included in '.c' files
as well as here. It seems reasonable to say that a header that might
have been C becomes C++ if it includes a C++ header; and lmi has no '.h'
headers.