Improve documentation
Don't refer to std::pow() in comments within files that don't use it
directly, so that `git grep -l '\<std::pow\>'` finds only actual
invocations. In many cases, nonstd::power() or a combination of
expm1() and log1p() is more accurate, and using std::pow() instead is
therefore a defect.