Improve documentation
In comments, don't qualify function names with 'std::' where a
nonstandard implementation might be substituted, rendering the
comment incorrect.

State the transformation
  (1+i)^n - 1 <-> expm1(log1p(i) * n)
OAOO, instead of repeating it several times.

pow() is both slower and less accurate than the combination of
expm1() and log1p(): not "less inaccurate" as previously stated.