For immediate reversion: binary exponentiation
Both SGI's power() and Knuth's Algorithm A in section 4.6.3 are
presumably well tested, and SGI's documentation implies that they
are equivalent, but that's hard to see at a glance.

Knuth calls this "right-to-left binary method for exponentiation".
The SGI documentation calls this the "Russian Peasant" algorithm,
but that's really something different (although related).