Prefer nonstd::power() to std::pow() in a particular case
The exponent in this case is always an integer in {1, 3, 6, 12}.
For x86_64, pc-linux-gnu and MinGW-w64 yielded different results
when std::pow() was used; nonstd::power() makes results identical.

The number of multiplications is low, so there's little opportunity
for error to accumulate. And, because the exponent is lower than 15,
right-to-left binary exponentiation uses the optimal addition chain,
so the number of multiplications is the lowest possible.

Although 'i' may be small (so that 'v' is close to unity), 'p' is
likely to be large, and therefore 'vp' is generally outside the range
where using
  vp^N <=> 1 + expm1(log1p(vp - 1) * N)
would be significantly more accurate.