Prepare to refactor
It is necessary to determine which targets can be built in any given
context (e.g., build_type, architecture). As the tangled conditionals
demonstrate, it is difficult to build this list additively. Experience
with unit-test targets in 'objects.make' shows that it is easier to
build a list of conditional exclusions. Accordingly, adapted that
exclusionary framework, intending to replace "add-unless" conditionals
with "exclude-if" ones in later commits--one by one, because a single
big-bang revision might easily introduce new defects, and also because
the present conditionals are not free of defects themselves.