Remove an unuseful option
Few $UBSAN_OPTIONS settings are generally useful--perhaps only one.

- UBSAN_OPTIONS=print_stacktrace=1 is generally useful. Without it, only
a one-line diagnostic is printed, which does not indicate which program
triggered the UB.

- UBSAN_OPTIONS=print_summary=1 seems merely to print "SUMMARY" and copy
a one-line diagnostic that would be printed anyway.

- UBSAN_OPTIONS=log_path="some_path" feels like a slapdash afterthought.
If "some_path" has a terminal '/', then output is silently discarded.
Without a terminal '/', a log is written, but with a randomized name
like "some_path.240247" that isn't available to scripts. The only solid
technique for processing UBSan output in scripts is to intercept stderr.