Automate UBSan unit tests
Adding UBSan support more than doubles the size of this script;
arguably, once it's copied to 'nychthemeral_test.sh', it should be
removed here. However, at least for the moment, this commit serves
to highlight the somewhat recherché technique, for which see:
  http://www.softpanorama.org/Tools/tee.shtml

It just so happens that UBSan writes its diagnostics to stderr,
which lmi's unit tests almost never use.