Specify .exe extension with paths passed to Wine
Since Wine 7, full paths to the executables without any extension don't
work any longer and result in a cryptic error message of the form

	wine: failed to open "/full/path/to/program": c0000135

So use the extension when running the unit tests: this is harmless for
all Wine versions (and was already done for e.g. test_coding_rules
before), and is required with Wine 7.