Set $EXEEXT more reliably for PETE
$EXEEXT had been set in a fragile way that is no longer correct.
Removed that code sedimentation, along with the invocation of
'config.guess' whose result is otherwise not used here. Instead,
transplanted a minimal few lines from 'set_toolchain.sh' into
'rebuild_pete.sh' to accomplish this one purpose.

Importing the heavyweight 'transume_toolchain.sh' mechanism from
'GNUmakefile' wouldn't make sense: the goal there is to make it
impossible to use a makefile incorrectly, but here it's okay if
incorrect use with msw produces an error, as this makefile is to
be run only by an expert or by the nychthemeral test and in either
case not without regard to the rebuild script obviously provided.