Suggest a workaround for a potential problem
Overriding $(test_dir) in a command such as this:
  make $coefficiency test_dir=/opt/lmi/test_clang system_test
causes an (empty) directory to be created, which leads to failure with
a "No testdecks" message.

Elaborated that message, suggesting a (manual) workaround.

Alternative not pursued: instead of echoing a command, run it. That
would not always be correct: the test and touchstone directories may
deliberately contain different sets of files, or different versions of
the same files. This problem occurs rarely, and only when $(test_dir)
is deliberately overridden; but there may be no reason ever to do that,
because all architectures supported now, or soon:
  LMI_COMPILER=gcc   ; LMI_TRIPLET=x86_64-pc-linux-gnu
  LMI_COMPILER=clang ; LMI_TRIPLET=x86_64-pc-linux-gnu
  LMI_COMPILER=gcc   ; LMI_TRIPLET=x86_64-w64-mingw32
produce exactly the same output for all 1485 tests.