Move a definition
The variable $(treat_warnings_as_errors) is of course not actually
necessary. It exists only to obstruct any temptation to alphabetize
'-pedantic-errors' and '-Werror' with other "warnings", where they
would be harder to notice. Unlike the others, these two are not
warnings as such: rather, they're options that make warnings fatal.

This is clearest when the variable is defined immediately before use.