Simplify
The code to decide whether '-fomit-frame-pointer' must be used with
MinGW-w64 is too complicated and confusing. Because ".0" is appended
to $(gcc_version), it can never match "10" exactly.

The debian-testing ("bookworm") MinGW-w64 gcc version is still 10.0.0
after all this time, and there's no newer version in "sid" today. It
seems unlikely ever to match 10.2.1 exactly: e.g., it may advance to
10.4 in a single jump.

Thus, expunged the innermost condition: it's unlikely ever to be
helpful, and it's confusing.