Fix defect introduced 20190327T2239Z: incomplete override
In 2019-02, $(tutelary_flag) was used to force $(product_file_flags) to
override all other options, for product files only. That was the only
way $(tutelary_flag) was used at the time. The overrides included
'-fno-omit-frame-pointer', which was universally used at that time.

Commit 376e4dfe7c4d4 of 20190327T2239Z used $(tutelary_flag) to force
'-fomit-frame-pointer' due to a compiler error. The intention was to do
that in every invocation of MinGW-w64-g++, but the $(product_file_flags)
then overrode that override, forcing the old '-fno-omit-frame-pointer'
behavior. This happened not to cause an ICE, so it was harmless; but if
a later product change happened to cause an ICE when compiling product
files, it might have been difficult to track down the reason.

Fixed the proximate cause of that potential problem by removing the
'-fno-omit-frame-pointer' option from $(product_file_flags). The
ultimate cause is that $(tutelary_flag) has been used inappropriately;
that will be fixed soon.