Resolve a latent makefile issue
A recently-added target-specific value of $(debug_flag) failed to
propagate through immediately-expanded variables. See:
  https://lists.nongnu.org/archive/html/lmi/2022-07/msg00005.html
Changed the expansion of the higher-level variables to recursive.

The issue is only latent because everything "works" without this
revision. The reason is that the '-fno-var-tracking-assignments'
option that was required with gcc-3.4.5 is no longer necessary.