Tidy up a couple of fardel commands
When no other command is echoed, 'printf' shouldn't be, either. Echoing
this line in particular would show JDNs, which aren't meant to be read
by humans.

Used '--auto-compress' to avoid specifying 'bzip2' compression twice.

Incidentally, compressed sizes are:
  70113052 /opt/lmi/fardels/lmi-20220704T2048Z.tar.bz2
  47264416 /opt/lmi/fardels/lmi-20220704T2048Z.tar.xz
  71679148 /opt/lmi/fardels/lmi-20220704T2048Z.zip

Clearly 'xz' compresses better, but 'bzip2' compression will continue
to be used in a corporate msw environment where tarballs are extracted
by a tangle of ancient scripts that it is better not to touch.

The '.zip' format is used for publication on savannah because it is
conventional for msw, but not for internal distribution because it is
not necessarily handled correctly in corporations, where many more files
are subject to md5 validation and "correcting" lmi's '\n' EOL indicators
causes validation to fail.