Explicitly enable '-Wpedantic' for gcc
Commit 1a2c6b496fbf1 of 20051031T0342Z replaced three flags with two:
  -  -pedantic \
  -  -pedantic-errors \
  -  -Werror \
  +treat_warnings_as_errors := -pedantic-errors -Werror
evidently in the belief that '-pedantic-errors' implies '-pedantic'.
The gcc documentation suggests otherwise:

  https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html
| -pedantic-errors ... is not equivalent to -Werror=pedantic, since
| there are errors enabled by this option and not enabled by the
| latter and vice versa. ...
| -Werror=switch turns the warnings controlled by -Wswitch into errors.
| ... -Werror=foo automatically implies -Wfoo.

It says, therefore, that there are errors enabled by '-Wpedantic' that
are not enabled by '-pedantic-errors', even though
  make show_overlooked_cxx_warnings
suggests otherwise. Added '-Wpedantic' anyway because it's harmless even
if redundant.

The documentation lists '-Wpedantic' and '-pedantic' together, with a
common definition, so those two must be equivalent. Apparently the '-W'
alias was introduced in gcc-4.8 or so.