When in doubt, prefer the Rule of Five to the Rule of Zero
The Rule of Zero is fine for trivial classes, but few if any of these
classes are trivial. See:
  https://lists.nongnu.org/archive/html/lmi/2022-07/msg00028.html

This revision reverts the deletion of dtors, and adds the four other
special member functions as "= delete" if possible or "= default" if
necessary for compiling and linking. Inspecting the haphazard outcomes
suggests that this is not a good strategy, though the exercise was
illuminating.

In class product_data, the copy ctor and copy assignment operator had
been private, with the former implemented out of line and the latter
deleted. The deleted member needn't be private. Perhaps the copy ctor
should be private, or perhaps not; that will be considered next, but
for now it's public.