Revert recent changes to one class's special member functions
The copy ctor is implemented, but private (and is now used only in a
unit test), while the copy assignment operator is deleted, and private.
This seems a bit unusual, but there's nothing wrong with it, and in
particular it doesn't depend on any deprecated behavior, so it's not
fated to become wrong. There was no really good reason to change it.

Comments indicating that a function is implemented out of line do not
say anything that is not already clear enough.

See:
  https://lists.nongnu.org/archive/html/lmi/2022-07/msg00032.html