Remove comments that have become untrue
The dtor must be declared explicitly because it must be virtual.
Declaring a dtor, even with "= default", prevents other special member
functions from being declared implicitly. (Not yet, for copy functions,
whose implicit declarations are only deprecated for now; but we're
designing for the future, as if what is deprecated today had already
been removed from the language.) Therefore, for them to be available
at all, they must all be declared explicitly. The removed comment tells
us that it is appropriate to default all of them; having done so, there
is no longer any comment to be written.