Make a different virtual pure
When at least one virtual must be declared pure because a base class
shouldn't be instantiated directly, it is conventional to put the
pure-specifier on the dtor. This commit challenges that convention.

C++20 says [class.abstract/2]: "A function declaration cannot provide
both a pure-specifier and a definition." Formerly, the dtor was
declared as pure, and elsewhere defined as defaulted, with comments
in both places. Code that requires comments is worse than code that
does not. Moving the pure-specifier made comments unnecessary. This
is therefore a pure improvement.

Forwent the usual space before "0" because it would have made the line
81 characters wide. To avoid writing a line wider than 80 characters,
in a file where that limit is otherwise respected except for the
copyright dates, is more important than writing that customary space.