Move semantics, nascent
This commit subsumes the soon-to-be-deleted odd/move_semantics branch.

It is conventional to place a header's implementation details in
namespace "detail", in order to mark them as notionally private. Here,
namespace "smf_mechanics" is used for the same purpose. For headers
that do not enclose all their contents in a namespace such as "lmi",
this seems preferable, to guard against collision of "details" in
different headers.