Experimentally perform specamt strategies during solves
This may or may not be a good idea. It does seem reasonable enough to
ask for a premium solve optionally to set specamt to corridor. However,
it wouldn't be unreasonable to disallow specamt strategies that depend
on premium when premium is being solved for. This commit and the ones
immediately preceding it add considerable complexity, and might
introduce subtle defects. Is this worth the trouble?