Augment documentation
Solving for premium to yield a target NAAR, and simultaneously
determining specified amount as a function of premium, has been seen
not to work well without a deep redesign of the solve code. Can such
specified-amount strategies therefore be grayed out when solving for
premium? No, not without redesigning the input-sequence classes, which
would take too much work for too little benefit.