Replace nongnu with llvm unwind library
* compiler_gcc_warnings.make: Prevent unit tests from failing. See:
  https://lists.nongnu.org/archive/html/lmi/2022-10/msg00006.html

* lmi_setup_20.sh: Install debian packages 'libc++-dev' (which replaces
'libunwind') and 'libc++abi-dev' (which is required for building
unwindable code).

* unwind.cpp: Remove UNW_LOCAL_ONLY macro, which would otherwise be
flagged as unused. The old HP library was intended to support remote
as well as local unwinding--see:
  https://lists.llvm.org/pipermail/cfe-dev/2016-September/050649.html
but the llvm library doesn't use that macro because it doesn't implement
remote unwinding.

* workhorse.make: Add the llvm library's include directory to the
'-isystem' include path (to suppress compiler warnings for its
contents). Do this unconditionally, for simplicity, even though msw
builds don't use it. It would seem preferable to include the principal
header as <libunwind/libunwind.h>, but that doesn't work--see:
  https://lists.nongnu.org/archive/html/lmi/2022-09/msg00011.html