Work around a libstdc++ issue less brutally
The <regex> header in libstdc++ is not UBSAN-clean when used with llvm's
libunwind. Inhibiting '-Wmaybe-uninitialized' for all unit tests that
use UBSAN, as in the last commit, was quite brutal. Instead, this commit
inhibits that warning for <regex> only, using a pragma that affects gcc
only...but without regard to UBSAN. This is not ideal, but a proposal to
add a builtin __SANITIZE_UNDEFINED__ macro:
  https://gcc.gnu.org/legacy-ml/gcc-patches/2019-09/msg01286.html
was rejected; and changing the build system to implement our own is too
much wasted work, as our time would be better spent repairing <regex>,
but that's silly because libstdc++'s implementation has such woeful
performance that it needs a wholesale replacement.

Similarly, it's hardly worthwhile to extend this "improved" workaround
to the other three object files in $(wno_maybe_uninitialized_objects),
e.g., by creating a new "regex_pragmatized" header to replace <regex>,
though it is worthwhile to avoid excluding all object files dependent on
any unit-test target.