Record speed measurements
Upgraded from wine-6.0.3 to wine-7.0 .

Compared to a777af49 (using only the "least" column for robustness):
  clang_gnu64: 3% faster
  gcc_gnu64: about the same for no-solve scenarios; else 20-30% slower
  gcc_msw64: +/- 2-3%; worst change is for solves, though

As the last several commits show, the remarkable speed regression for
gcc_gnu64 is a result of upgrading from gcc version eleven to twelve.
It seems likely that this represents a defect in version twelve...which
conceivably might be corrected in due time, and before the msw version
reaches version twelve.