Record gcc version eleven pc-linux-gnu speed for commit 5a4fcaed7
Nothing remarkable here: this commit serves only to establish a base for
comparison, to show a speed regression with gcc version twelve. In order
to reproduce these historical results, it is imperative to run commands
such as
  # apt-get install libunwind-dev
  # apt-get install libc++-dev
to install a contemporary unwind library--each of those two commands
undoes what the other does in that respect.

Results using commit 5a4fcaed7 of 2022-08-18T16:22:15.

$git --no-pager log -1 --oneline
5a4fcaed7 (HEAD) Begin adapting to gcc-12

$gcc -dumpfullversion
11.2.0