Use clang's preprocessor
The preprocessor, traditionally $CPP, is not explicitly used by any lmi
makefile; therefore, it is unspecified for clang.

The technique lmi uses to generate autodependency files relies on the
  -MMD -MP -MT -MF
options of the gcc preprocessor, which is invoked as $GNU_CPP even with
compilers incompatible with gcc that lack those options; but clang's
preprocessor deliberately supports those options for compatibility with
gcc's; therefore, it is used for that purpose.

Clang provides a separate 'clang-cpp-*' preprocessor, but at least in
debian it's versioned: e.g., 'clang-cpp-13' exists today in 'bookworm',
but there's no generic 'clang-cpp'. Therefore, '$(CXX) -E' is used here,
on the assumption that the C++ compiler's preprocessor works correctly
for C source files (at least for autodependencies).