Allow 'tee /dev/tty' to fail
Building lmi on a corporate server has become difficult because the
corporate VPN often disconnects, sending a hangup signal to the server
even if the VPN automatically reconnects immediately. Building with:
  nohup gwc/redhat_rebuild.sh & disown
would seem to work, except that /dev/tty no longer exists when the
terminal is closed or disconnected, causing writes to that nonexistent
device to fail.

No more elaborate method seems warranted. The intention is only to
write occasional messages to the terminal when the output of that
long-running script has been redirected, when it's run on a personal
workstation. On this corporate server, that's just not feasible.