Revert "Add temporary statements to debug a script"
This reverts commit b6c641f9cc2c24208eb319b711ea17d738b5453e.

The diagnostic statements removed here were correctly written to file
'debian-log_*'. The conditional filtering on `tty -s` affects only
output that would be written to the screen, not output written to the
nychthemeral test's log directory, in this pipeline:

  | tee "$log_dir"/nychthemeral_test | sed -e "$nychthemeral_clutter"
    ^^^^^^^^^^^nonfiltered^^^^^^^^^^   ^^^^^^^^^^^filtered^^^^^^^^^^^

Thus, when a nychthemeral test is run in the normal way, the presence
of any output other than the architecture-naming progress indicators
indicates an actual error; but when it's run as
  nohup /wherever/opt/lmi/src/lmi/gwc/redhat_rebuild.sh & disown; exit
nothing can we written to the screen because there is no tty. Either
way, unwanted 'wine' messages appear in "$log_dir"/nychthemeral_test;
no attempt is currently made to filter them because they're pervasive
yet 'wine' provides no fine-grained facility to suppress them--see:
  https://lists.nongnu.org/archive/html/lmi/2023-01/msg00019.html
and suppressing whole categories of important messages with $WINEDEBUG
would create a risk that some significant message might be missed.