Show message box safely in wx warning_alert() implementation
This makes this function safe to call even before the GUI is
initialized.

While this function can't simply use wxSafeShowMessage() because it
doesn't have quite the semantics we want, it can use a lower-level
function which only shows the message box without using stderr.

Note that we could also do

    if(wxSafeShowMessage("Warning", s))
        {
            std::cerr << "Warning: " << s << std::endl;
        }

which would have almost the same effect and have the advantage of being
simpler, but we prefer to use this version in order to log the message
to stderr before showing the message box instead of doing it after
showing it and so behave in exactly the same way as the current code.