Avoid MSVC warnings in bourn_cast unit test
These warnings warn about the problems expected in this code test, with,
perhaps, the only exception of a warning in the assignment to "min"
which seems to be bogus as the compiler complains about unsafe use of
"bool" and "long double" with operator==() which is not present in this
line, so suppress them using the same pragma-based approach that was
already used for the other compilers.

No real changes.